EDIT: Refactored - move scripts into folder scripts
