+++ test/models/test_ensemble.py
@@ -8,7 +8,7 @@ import unittest
 
 import numpy as np
 
-import autosklearn.ensemble_script
+import scripts.ensemble_script
 from autosklearn.constants import *
 
 N_TEST_RUNS = 10
@@ -35,7 +35,7 @@ class Test(unittest.TestCase):
         metric = 'f1_metric'
         task_type = MULTICLASS_CLASSIFICATION
 
-        weights = autosklearn.ensemble_script.weighted_ensemble(
+        weights = scripts.ensemble_script.weighted_ensemble(
             predictions, true_labels, task_type, metric, weights)
 
         self.assertEqual(weights.shape[0], self.n_models)
@@ -50,7 +50,7 @@ class Test(unittest.TestCase):
 
         pred = np.random.rand(self.n_models, self.n_points, self.n_classes)
         w = np.ones([self.n_models]) * (1.0 / float(self.n_models))
-        p_hat = autosklearn.ensemble_script.ensemble_prediction(pred, w)
+        p_hat = scripts.ensemble_script.ensemble_prediction(pred, w)
         p = pred.mean(axis=0)
 
         # First test case
@@ -61,7 +61,7 @@ class Test(unittest.TestCase):
         # Second test case
         w = np.zeros([self.n_models])
         w[0] = 1.0
-        p_hat = autosklearn.ensemble_script.ensemble_prediction(pred, w)
+        p_hat = scripts.ensemble_script.ensemble_prediction(pred, w)
         for i in range(self.n_points):
             for j in range(self.n_classes):
                 self.assertAlmostEqual(pred[0, i, j], p_hat[i, j])
@@ -78,7 +78,7 @@ class Test(unittest.TestCase):
         metric = 'f1_metric'
         task_type = MULTICLASS_CLASSIFICATION
 
-        err = autosklearn.ensemble_script.weighted_ensemble_error(
+        err = scripts.ensemble_script.weighted_ensemble_error(
             weights, predictions, true_labels, metric, task_type)
 
         self.assertAlmostEqual(err, 1.0, delta=0.3)
